Added PUT POST employee

+++ spring-backend/src/main/java/manago/com/restbackend/service/impl/EmployeeServiceImpl.java
@@ -1,8 +1,12 @@
 package manago.com.restbackend.service.impl;
 
+import manago.com.restbackend.exception.model.ErrorMessages;
+import manago.com.restbackend.model.Employee;
 import manago.com.restbackend.repository.EmployeeRepository;
 import manago.com.restbackend.repository.TeamRepository;
+import manago.com.restbackend.repository.UserRepository;
 import manago.com.restbackend.service.EmployeeService;
+import manago.com.restbackend.shared.request.EmployeeRequest;
 import manago.com.restbackend.shared.response.EmployeeResponse;
 import manago.com.restbackend.util.ManagoMapper;
 import org.springframework.stereotype.Service;
@@ -15,11 +19,13 @@ public class EmployeeServiceImpl implements EmployeeService {
 
     private EmployeeRepository employeeRepository;
     private TeamRepository teamRepository;
+    private UserRepository userRepository;
     private ManagoMapper mapper;
 
-    public EmployeeServiceImpl(EmployeeRepository employeeRepository, TeamRepository teamRepository, ManagoMapper mapper) {
+    public EmployeeServiceImpl(EmployeeRepository employeeRepository, TeamRepository teamRepository, UserRepository userRepository, ManagoMapper mapper) {
         this.employeeRepository = employeeRepository;
         this.teamRepository = teamRepository;
+        this.userRepository = userRepository;
         this.mapper = mapper;
     }
 
@@ -33,13 +39,48 @@ public class EmployeeServiceImpl implements EmployeeService {
     }
 
     public EmployeeResponse one(Long id) {
-        return mapper.employeeToEmployeeResponse(employeeRepository.findByEmployeeId(id));
+        Employee employee = employeeRepository.findByEmployeeId(id);
+        if (employee == null)
+            throw new RuntimeException(ErrorMessages.RECORD_NOT_FOUND.getErrorMessage());
+        else return mapper.employeeToEmployeeResponse(employee);
+    }
+
+    public EmployeeResponse updateEmployee(Long id, EmployeeRequest request) {
+        Employee employee = employeeRepository.findByEmployeeId(id);
+        if (employee == null)
+            throw new RuntimeException(ErrorMessages.RECORD_NOT_FOUND.getErrorMessage());
+        else {
+            employee.setEmploymentType(request.getEmploymentType());
+            employee.setFirstName(request.getFirstName());
+            employee.setLastName(request.getLastName());
+            employee.setSalary(request.getSalary());
+
+            employeeRepository.save(employee);
+            return mapper.employeeToEmployeeResponse(employee);
+        }
+    }
+
+    public EmployeeResponse createEmployee(EmployeeRequest request) {
+        Employee employee = mapper.employeeRequestToEmployee(request);
+        employeeRepository.save(employee);
+        return mapper.employeeToEmployeeResponse(employee);
     }
 
     public void delete(Long id) {
-        teamRepository.findAll()
-                .stream()
-                .forEach(team -> team.deleteEmployee(employeeRepository.findByEmployeeId(id)));
-        employeeRepository.deleteById(id);
+        if (employeeRepository.findByEmployeeId(id) == null)
+            throw new RuntimeException(ErrorMessages.RECORD_NOT_DELETED.getErrorMessage());
+        else {
+            //Delete all information about employee in teams
+            teamRepository.findAll()
+                    .forEach(team -> team.deleteEmployee(employeeRepository.findByEmployeeId(id)));
+
+            //Delete all information about employee in users
+            userRepository.findAllByEmployee(employeeRepository.findByEmployeeId(id))
+                    .forEach(user -> {
+                        user.setEmployee(null);
+                        userRepository.save(user);
+                    });
+            employeeRepository.deleteById(id);
+        }
     }
 }
